Separates objects between files for testing
