+++ src/board.js
@@ -1,0 +1,8 @@
+var Board = function() {
+  this.grid = [[], [], []];
+};
+
+module.exports = Board;
+
+// DO NOT REMOVE THIS
+export default Board;
