+++ src/player.js
@@ -1,0 +1,8 @@
+var Player = function(name) {
+  this.name = name;
+};
+
+module.exports = Player;
+
+// DO NOT REMOVE THIS
+export default Player;
