+++ src/tictactoe.js
@@ -1,15 +1,10 @@
+import Board from 'board';
+import Player from 'player';
+
 var TicTacToe = function() {
   this.board = new Board();
   this.player1 = new Player();
   this.player2 = new Player();
-};
-
-var Board = function() {
-  this.grid = [[], [], []];
-};
-
-var Player = function(name = "Testy") {
-  this.name = name;
 };
 
 module.exports = TicTacToe;
